# feat(salts): optimize salts storage