```
name            old time/op    new time/op    delta
Salts_Get-12       132ns ± 1%      15ns ± 1%   -88.87%  (p=0.000 n=10+10)
Salts_Store-12    90.6ns ± 7%   139.0ns ± 1%   +53.43%  (p=0.000 n=10+9)

name            old alloc/op   new alloc/op   delta
Salts_Get-12       96.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Salts_Store-12     0.00B         24.00B ± 0%     +Inf%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
Salts_Get-12        3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Salts_Store-12      0.00           1.00 ± 0%     +Inf%  (p=0.000 n=10+10
```